Fix alldata parsing to allow files >4GB

+++ MArchiveBatchTool/MArchive/AllDataPacker.cs
@@ -106,7 +106,7 @@ namespace GMWare.M2.MArchive
                 ArchiveV1 archive = new ArchiveV1();
                 archive.ObjectType = "archive";
                 archive.Version = 1.0f;
-                archive.FileInfo = new Dictionary<string, List<int>>();
+                archive.FileInfo = new Dictionary<string, List<long>>();
                 folderPath = Path.GetFullPath(folderPath);
 
                 foreach (var file in Directory.GetFiles(folderPath, "*", SearchOption.AllDirectories))
@@ -117,7 +117,7 @@ namespace GMWare.M2.MArchive
                     using (FileStream fs = File.OpenRead(file))
                     {
                         fs.CopyTo(packStream);
-                        archive.FileInfo.Add(key.Replace('\\', '/'), new List<int>() { (int)currPos, (int)fs.Length });
+                        archive.FileInfo.Add(key.Replace('\\', '/'), new List<long>() { currPos, fs.Length });
                     }
                     var targetLength = (packStream.Position + ALIGNMENT - 1) / ALIGNMENT * ALIGNMENT;
                     byte[] alignBytes = new byte[targetLength - packStream.Position];
