+++ MArchiveBatchTool/Utils.cs
@@ -37,11 +37,11 @@ namespace GMWare.M2
         /// <param name="output">The stream to copy to.</param>
         /// <param name="count">The number of bytes to copy.</param>
         // Modified from https://stackoverflow.com/a/230141/1180879
-        public static void CopyStream(Stream input, Stream output, int count)
+        public static void CopyStream(Stream input, Stream output, long count)
         {
             byte[] buffer = new byte[81920];
             int read;
-            while (count > 0 && (read = input.Read(buffer, 0, Math.Min(buffer.Length, count))) > 0)
+            while (count > 0 && (read = input.Read(buffer, 0, (int)Math.Min(buffer.Length, count))) > 0)
             {
                 output.Write(buffer, 0, read);
                 count -= read;
